Stack walking.
The stack can now be walked to look at live values. To enable this, the stack layout needed to change. At RBP-8, the current function address is stored, instead of the first argument/register.
Also, the liveness analysis have been changed such that local references are alive the whole function.